app/requirements.txt
cachetools==5.2.0
coverage-badge==1.1.0
fastapi==0.75.1
google-api-core==2.8.2
google-auth==2.8.0
google-cloud-core==2.3.1
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.3
Jinja2==3.1.1
joblib==1.1.0
lightgbm==3.3.2
numpy==1.22.3
pandas==1.4.2
protobuf==4.21.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytest==7.1.1
requests==2.27.1
rsa==4.8
scikit-learn==1.0.2
Shapely==1.8.1.post1
uvicorn[standard]==0.17.6
xgboost==1.5.2
